scheduler/requirements.txt
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pika==1.3.2
pymongo==4.15.1
Werkzeug==3.1.3
python-dotenv==1.2.1
